Update the running to use the BasText files; invoking scripts.
The code now loads the files with TEXTLOAD and continues execution
by running the RebindKeys script.

The module is no longer loaded from memory after the rebinding (the
building of the module shouldn't execute it; that's a separate thing
if you want it).

There's no need to kill ShellCLI (and the JFPatch-as-a-service
environment doesn't have one).