Refactor the CI run to be smaller and more reusable.
The steps are now grouped together so that there's a 'build it'
step, a 'show what happened' step, and a 'put the archive in a place
that we can archive it' step. The latter step uses a VersionNum file
to give the version of the module which is appended to the archive
we produce, like most software releases do. The VersionNum is used
in the module to ensure that we keep these consistent.